Processing Used In Manuscript
This commit contains the data extraction and processing code that was used in the manuscript. There is only an implementation for data found in two consecutive scans. Although more consecutive data points would lead to greater confidence in the extrapolated precursor m/z determination, the time penalty is too great and therefore impractical for DIA on a tribrid mass spectrometer!